refactor(ui): component refactoring
- moved GoBackButton to a separate component
- moved CLIEquivalent to a separate component
- moved KopiaTable to components directory
