+++ src/components/KopiaTable.jsx
