+++ src/pages/Repository.jsx
@@ -9,7 +9,7 @@ import InputGroup from "react-bootstrap/InputGroup";
 import Spinner from "react-bootstrap/Spinner";
 import { handleChange } from "../forms";
 import { SetupRepository } from "../components/SetupRepository";
-import { CLIEquivalent } from "../utils/uiutil";
+import { CLIEquivalent } from "../components/CLIEquivalent";
 import { cancelTask } from "../utils/taskutil";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCheck, faChevronCircleDown, faChevronCircleUp, faWindowClose } from "@fortawesome/free-solid-svg-icons";
