+++ src/utils/uiutil.jsx
@@ -1,12 +1,7 @@
-import { faChevronLeft, faCopy, faExclamationTriangle, faTerminal } from "@fortawesome/free-solid-svg-icons";
+import { faExclamationTriangle } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import axios from "axios";
-import React, { useState } from "react";
-import Button from "react-bootstrap/Button";
-import FormControl from "react-bootstrap/FormControl";
-import InputGroup from "react-bootstrap/InputGroup";
-import { useNavigate, Link } from "react-router-dom";
-import PropTypes from "prop-types";
+import React from "react";
+import { Link } from "react-router-dom";
 
 // locale to use for number formatting (undefined would use default locale, but we stick to EN for now)
 const locale = "en-US";
@@ -271,16 +266,6 @@ export function formatDuration(from, to, useMultipleUnits = false) {
   return formatMilliseconds(ms, useMultipleUnits);
 }
 
-export function GoBackButton() {
-  const navigate = useNavigate();
-
-  return (
-    <Button size="sm" variant="warning" onClick={() => navigate(-1)}>
-      <FontAwesomeIcon icon={faChevronLeft} /> Return{" "}
-    </Button>
-  );
-}
-
 export function PolicyTypeName(s) {
   if (!s.host && !s.userName) {
     return "Global Policy";
@@ -385,68 +370,6 @@ export function errorAlert(err, prefix) {
     alert(prefix + JSON.stringify(err));
   }
 }
-
-export function CLIEquivalent(props) {
-  let [visible, setVisible] = useState(false);
-  let [cliInfo, setCLIInfo] = useState({});
-
-  if (visible && !cliInfo.executable) {
-    axios
-      .get("/api/v1/cli")
-      .then((result) => {
-        setCLIInfo(result.data);
-      })
-      .catch((_error) => {});
-  }
-
-  const ref = React.createRef();
-
-  function copyToClibopard() {
-    const el = ref.current;
-    if (!el) {
-      return;
-    }
-
-    el.select();
-    el.setSelectionRange(0, 99999);
-
-    document.execCommand("copy");
-  }
-
-  return (
-    <>
-      <InputGroup size="sm">
-        <Button
-          data-testid="show-cli-button"
-          size="sm"
-          title="Click to show CLI equivalent"
-          variant="submit"
-          onClick={() => setVisible(!visible)}
-        >
-          <FontAwesomeIcon size="sm" icon={faTerminal} />
-        </Button>
-        {visible && (
-          <Button size="sm" variant="success" title="Copy to clipboard" onClick={copyToClibopard}>
-            <FontAwesomeIcon size="sm" icon={faCopy} />
-          </Button>
-        )}
-        {visible && (
-          <FormControl
-            size="sm"
-            ref={ref}
-            className="cli-equivalent"
-            readOnly={true}
-            value={`${cliInfo.executable} ${props.command}`}
-          />
-        )}
-      </InputGroup>
-    </>
-  );
-}
-
-CLIEquivalent.propTypes = {
-  command: PropTypes.string.isRequired,
-};
 
 export function toAlgorithmOption(x, defaultID) {
   let text = x.id;
