+++ tests/components/GoBackButton.test.jsx
@@ -1,0 +1,44 @@
+import React from "react";
+import { render, screen, fireEvent } from "@testing-library/react";
+import { BrowserRouter } from "react-router-dom";
+import { vi } from "vitest";
+import "@testing-library/jest-dom";
+import { GoBackButton } from "../../src/components/GoBackButton";
+
+// Mock react-router-dom navigate
+const mockNavigate = vi.fn();
+vi.mock("react-router-dom", async () => {
+  const actual = await vi.importActual("react-router-dom");
+  return {
+    ...actual,
+    useNavigate: () => mockNavigate,
+  };
+});
+
+// Helper to render components with router
+const renderWithRouter = (component) => {
+  return render(<BrowserRouter>{component}</BrowserRouter>);
+};
+
+describe("GoBackButton", () => {
+  beforeEach(() => {
+    mockNavigate.mockClear();
+  });
+
+  it("renders with correct text and icon", () => {
+    renderWithRouter(<GoBackButton />);
+
+    const button = screen.getByRole("button");
+    expect(button).toBeInTheDocument();
+    expect(button).toHaveTextContent("Return");
+  });
+
+  it("calls navigate(-1) when clicked", () => {
+    renderWithRouter(<GoBackButton />);
+
+    const button = screen.getByRole("button");
+    fireEvent.click(button);
+
+    expect(mockNavigate).toHaveBeenCalledWith(-1);
+  });
+});
